REST API 모음/웍샵10_2차/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.4
decorator==5.1.1
Django==3.2.7
django-extensions==3.1.5
django-seed==0.3.1
djangorestframework==3.13.1
executing==0.8.3
Faker==13.4.0
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toposort==1.7
traitlets==5.1.1
wcwidth==0.2.5
